chore(ui): finalize the Material 3 Expressive branch
Dead code: drop the unused Modifier.bounceClick (the file is now PressBounce.kt
with just the wired-in pressBounce) and the stale FilterChip/FilterChipDefaults
imports left in the five Library*Screen files after the toggle-chip swap.

Ratchets/tests: add ui-audit R25 (route content spinners through the shared
ZemerLoadingIndicator) and tighten four burned-down baseline entries; add
AvatarShapeIndexTest for the pure avatar-shape selection. The rest of the
expressive UI is Compose-only and not unit-testable without Robolectric.

Docs: refresh AGENTS.md and the regenerated code-derived docs for the full set
(loaders, wavy bars, toggle chips, carousel, cookie avatars, toggle pop, the
app-wide press bounce, and the high refresh rate toggle).
