+++ docs/reference/non-kotlin-files.md
@@ -11,7 +11,7 @@ Every tracked non-Kotlin path outside `docs/` is listed. Text files report line 
 | `.github/workflows/ui-audit.yml` | 50 lines | text `.yml` |
 | `.gitignore` | 117 lines | text `[none]` |
 | `.gitmodules` | 6 lines | text `[none]` |
-| `AGENTS.md` | 1452 lines | text `.md` |
+| `AGENTS.md` | 1460 lines | text `.md` |
 | `LICENSE` | 674 lines | text `[none]` |
 | `README.md` | 19 lines | text `.md` |
 | `app/.gitignore` | 1 lines | text `[none]` |
@@ -295,8 +295,8 @@ Every tracked non-Kotlin path outside `docs/` is listed. Text files report line 
 | `scripts/check-download-unification.sh` | 58 lines | text `.sh` |
 | `scripts/dead-resources-baseline.txt` | 9 lines | text `.txt` |
 | `scripts/telegram-chats.sh` | 38 lines | text `.sh` |
-| `scripts/ui-audit-baseline.tsv` | 24 lines | text `.tsv` |
-| `scripts/ui-audit.sh` | 235 lines | text `.sh` |
+| `scripts/ui-audit-baseline.tsv` | 20 lines | text `.tsv` |
+| `scripts/ui-audit.sh` | 241 lines | text `.sh` |
 | `scripts/ui-strings-scan.py` | 96 lines | text `.py` |
 | `settings.gradle.kts` | 56 lines | text `.kts`; plugins `org.gradle.toolchains.foojay-resolver-convention` |
 | `simpmusic/build.gradle.kts` | 15 lines | text `.kts`; plugins `kotlin.serialization, jvm` |
